upgrade deps

+++ src/caching_writer.rs
@@ -26,7 +26,7 @@ pub fn async_read_and_write_to_cache<'a>(
     let inp = Box::pin(inp);
     let mut zstd_writer = Some(ZstdEncoder::with_quality(
         Vec::new(),
-        async_compression::Level::Precise(compression_level as u32),
+        async_compression::Level::Precise(compression_level),
     ));
     let mut bytes_written = 0;
